requirements.txt
boto3==1.26.108
mangum==0.17.0
pytest==7.2.2
pytest-cov==4.0.0
reportlab==4.0.4
uvicorn==0.21.1
virtualenv==20.21.0
python-dotenv==1.0.0